Replaced mysql_ functions with mysqli_ functions
Since mysql and all mysql_ related functions are being deprecated in PHP
with the intent to being removed and as ZenCart is moving away from
mysql to mysqli being the new default, all references to
mysql_(anything) are being edited to appropriate mysqli functions
instead. A check is being done by adding a single define to
\YOUR_ADMIN\includes\extra_datafiles\super_order_defines.php by checking
the current version of ZenCart (which is defined globally). If it finds
that the current ZenCart version is 1.5.2 or lower, it will still
perform mysql and any mysql related function.