General fixes for checkbox related pages
This set of changes removes the need to refresh the page for a check all
button. Instead, it uses Javascript to check all checkboxes on a page by
adding the checkbox area to a div and using JS to cycle through and
check all the checkboxes. The also removes the $_GET['checked'] and
$checked related variables.