Removed MySQLi/MySQL switch
As suggested by lhungil, I removed the check for mysqli_ or mysqli
function. The `zen_db_perform` SHOULD be return something for
`$db->Insert_ID()` to use as this code was also prevalent in
`/includes/classes/order.php`. Also zen_db_delete no longer returns the
number of deleted fields as there is no reference to this value
whatsoever.